changed connectionKeeper
